% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp-class.R
\docType{class}
\name{gcp-class}
\alias{gcp-class}
\alias{.gcp}
\alias{gcp}
\title{GCP platform class}
\usage{
gcp()
}
\value{
An object of class \code{gcp}.
}
\description{
This class is used to represent the GCP platform.
}
\examples{
showClass("gcp")
}
