% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkflow-methods.R
\name{avworkflow-methods}
\alias{avworkflow-methods}
\title{AnVIL workflow methods}
\usage{
\S4method{avworkflow_jobs}{gcp}(
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  ...,
  platform = cloud_platform()
)
}
\value{
\code{avworkflow_jobs()} returns a \code{tibble}, sorted by
\code{submissionDate}, with columns
\itemize{
\item submissionId character() job identifier from the workflow runner.
\item submitter character() AnVIL user id of individual submitting the job.
\item submissionDate POSIXct() date (in local time zone) of job submission.
\item status character() job status, with values 'Accepted' 'Evaluating'
'Submitting' 'Submitted' 'Aborting' 'Aborted' 'Done'
\item succeeded integer() number of workflows succeeding.
\item failed integer() number of workflows failing.
}
}
\description{
Methods for working with AnVIL workflow execution.
\code{avworkflow_jobs()} returns a tibble summarizing submitted workflow jobs for
a namespace and name.
}
\examples{
if (
    gcloud_exists() && identical(avplatform_namespace(), "AnVILGCP") &&
    nzchar(avworkspace_name())
)
    ## from within AnVIL
    avworkflow_jobs()

}
